Physics Have been put

+++ Source/Engine/Renderer/Moedel.cpp
@@ -34,7 +34,7 @@ namespace Loki {
 		if (m_points.empty()) return;
 
 		renderer.SetColor(Color::ToInt(m_color.r), Color::ToInt(m_color.g), Color::ToInt(m_color.b), Color::ToInt(m_color.a));
-		for (int i = 0; i < m_points.size() - 1; i++) {
+		for (int i = 0; i < (int)m_points.size() - 1; i++) {
 			vec2 p1 = (m_points[i] * scale).Rotate(roatation) + position;
 			vec2 p2 = (m_points[i + 1] * scale).Rotate(roatation) + position;
 
@@ -50,7 +50,7 @@ namespace Loki {
 		mat3 mx = transform.GetMatrix();
 
 		renderer.SetColor(Color::ToInt(m_color.r), Color::ToInt(m_color.g), Color::ToInt(m_color.b), Color::ToInt(m_color.a));
-		for (int i = 0; i < m_points.size() - 1; i++) {
+		for (int i = 0; i < (int)m_points.size() - 1; i++) {
 			vec2 p1 = mx * m_points[i];
 			vec2 p2 = mx * m_points[i + 1];
